Update guides order
