--- docs/lib/mdx/index.ts
@@ -1,11 +1,0 @@
-// Templates
-export {
-	getTemplate,
-	getTemplateBreadcrumbs,
-	getTemplateDocsContent,
-	getTemplateSlugs,
-	getTemplateSubNavItems,
-	getTemplateList,
-	getTemplateNavLinks,
-} from './templates';
-export type { Template, TemplateList } from './templates';
