+++ docs/lib/mdxUtils.ts
@@ -9,14 +9,12 @@ import { slugify } from './slugify';
 
 const PKG_PATH = normalize(`${process.cwd()}/../packages/`);
 const RELEASE_PATH = normalize(`${process.cwd()}/../releases/`);
-const GUIDE_PATH = normalize(`${process.cwd()}/../guides/`);
 
 const pkgDocsPath = (slug: string) =>
 	normalize(`${PKG_PATH}/${slug}/README.md`);
 const pkgJsonPath = (slug: string) =>
 	normalize(`${PKG_PATH}/${slug}/package.json`);
 const releasePath = (slug: string) => normalize(`${RELEASE_PATH}/${slug}.mdx`);
-const guidePath = (slug: string) => normalize(`${GUIDE_PATH}/${slug}.mdx`);
 
 export function stripMdxExtension(filename: string) {
 	return filename.replace(/\.mdx?$/gi, '');
@@ -235,63 +233,3 @@ export type Release = Awaited<ReturnType<typeof getRelease>>;
 export type ReleaseList = Awaited<ReturnType<typeof getReleaseList>>;
 
 // Guides
-
-export async function getGuide(slug: string) {
-	const { content, data } = await getMarkdownData(guidePath(slug));
-	const source = await serializeMarkdown(content, data);
-
-	return {
-		slug,
-		source,
-		data,
-		title: (data.title ?? slug) as string,
-	};
-}
-
-export async function getGuideSlugs() {
-	const entries = await readdir(GUIDE_PATH, { withFileTypes: true });
-	return entries
-		.filter(
-			(entry) =>
-				!entry.name.startsWith('_') &&
-				!entry.name.startsWith('.') &&
-				!entry.name.startsWith('index') &&
-				entry.isFile()
-		)
-		.map((entry) => slugify(stripMdxExtension(entry.name)))
-		.sort()
-		.reverse();
-}
-
-export function getGuideList() {
-	return getGuideSlugs().then((slugs) =>
-		Promise.all(
-			slugs.map((slug) =>
-				getMarkdownData(guidePath(slug)).then(({ data }) => ({
-					title: (data?.title ?? slug) as string,
-					slug,
-				}))
-			)
-		)
-	);
-}
-
-function guideNavMetaData(
-	slug: string,
-	data: Awaited<ReturnType<typeof getMarkdownData>>['data']
-) {
-	return {
-		title: (data?.title ?? slug) as string,
-		slug,
-	};
-}
-
-export function getGuidesBreadcrumbs(slug: string) {
-	return getMarkdownData(guidePath(slug)).then(({ data }) => {
-		const meta = guideNavMetaData(slug, data);
-		return [{ href: '/guides', label: 'Guides' }, { label: meta.title }];
-	});
-}
-
-export type Guide = Awaited<ReturnType<typeof getGuide>>;
-export type GuideList = Awaited<ReturnType<typeof getGuideList>>;
