+++ docs/pages/guides/[slug].tsx
@@ -8,11 +8,12 @@ import {
 	getGuidesBreadcrumbs,
 	getGuideSlugs,
 	Guide,
-} from '../../lib/mdxUtils';
+} from '../../lib/mdx/guides';
 import { mdxComponents } from '../../components/utils';
 import { AppLayout } from '../../components/AppLayout';
 import { DocumentTitle } from '../../components/DocumentTitle';
 import { PageLayout } from '../../components/PageLayout';
+import { PageTitle } from '../../components/PageTitle';
 
 export default function Guides({
 	guide,
@@ -32,7 +33,7 @@ export default function Guides({
 					editPath={`/guides/${guide.slug}.mdx`}
 					breadcrumbs={breadcrumbs}
 				>
-					<H1>{guide.data.title}</H1>
+					<PageTitle title={guide.title} introduction={guide.description} />
 					<Body>
 						<MDXRemote {...guide.source} components={mdxComponents} />
 					</Body>
