+++ guides/forms.mdx
@@ -1,12 +1,10 @@
 ---
+order: 2
 title: Forms
+description: Web forms are one of the main points of interaction between a user and a web site or application. In this guide, we will be building a simple sign in form using two popular libraries react-hook-form and yup.
 ---
 
-Web forms are one of the main points of interaction between a user and a web site or application. Forms allow users to enter data, which is generally sent to a web server for processing and storage.
-
-In this guide, we will be building a simple sign in form using `react-hook-form` and `yup`. Fore more information about these two libraries, please refer to the [official documentation](https://react-hook-form.com/get-started#SchemaValidation).
-
-> The AgDS component library does not have any opinions for how form state should be handled. All of our form components have been designed to work with any form library.
+> Note: The AgDS component library does not have any opinions for how form state should be handled. All of our form components have been designed to work with any form library.
 
 ## 1. Install AgDS packages
 
@@ -21,7 +19,7 @@ yarn add @ag.ds-next/field @ag.ds-next/icon @ag.ds-next/text-input @ag.ds-next/b
 Using the `TextInput`, `FormStack` and `Button` components, we can quickly compose the user interface for the form.
 
 ```tsx
-import { Button } from '@ag.ds-next/button';
+import { Button, ButtonGroup } from '@ag.ds-next/button';
 import { FormStack } from '@ag.ds-next/form-stack';
 import { TextInput } from '@ag.ds-next/text-input';
 
@@ -31,9 +29,9 @@ function FormExampleSignIn() {
 			<FormStack>
 				<TextInput label="Email address" type="email" maxWidth="xl" required />
 				<TextInput label="Password" type="password" maxWidth="xl" required />
-				<div>
+				<ButtonGroup>
 					<Button type="submit">Sign in</Button>
-				</div>
+				</ButtonGroup>
 			</FormStack>
 		</form>
 	);
@@ -86,7 +84,7 @@ For more information about `react-hook-form`, please refer to the [official docu
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
-import { Button } from '@ag.ds-next/button';
+import { Button, ButtonGroup } from '@ag.ds-next/button';
 import { FormStack } from '@ag.ds-next/form-stack';
 import { TextInput } from '@ag.ds-next/text-input';
 
@@ -132,9 +130,9 @@ export function FormExampleSignIn() {
 					maxWidth="xl"
 					required
 				/>
-				<div>
+				<ButtonGroup>
 					<Button type="submit">Sign in</Button>
-				</div>
+				</ButtonGroup>
 			</FormStack>
 		</form>
 	);
@@ -149,7 +147,7 @@ By making use of the `invalid` and `message` props available to the `TextInput` 
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
-import { Button } from '@ag.ds-next/button';
+import { Button, ButtonGroup } from '@ag.ds-next/button';
 import { FormStack } from '@ag.ds-next/form-stack';
 import { TextInput } from '@ag.ds-next/text-input';
 
@@ -199,9 +197,9 @@ export function FormExampleSignIn() {
 					maxWidth="xl"
 					required
 				/>
-				<div>
+				<ButtonGroup>
 					<Button type="submit">Sign in</Button>
-				</div>
+				</ButtonGroup>
 			</FormStack>
 		</form>
 	);
