+++ guides/responsive-props.mdx
@@ -1,10 +1,8 @@
 ---
+order: 3
 title: Responsive props
+description: The Agriculture Design System provides first class support for mobile-first responsive styles. Instead of manually writing CSS media queries throughout your codebase, our primitive components allow you to provide arrays and objects to generate responsive styles in your project.
 ---
-
-The Agriculture Design System provides first class support for mobile-first responsive styles.
-
-Instead of manually writing CSS media queries throughout your codebase, our primitive components (`Box`/`Flex`/`Stack`) allow you to provide arrays and objects to generate responsive styles in your project.
 
 ## Breakpoints
